web/400-sql_injection/rit_directory/requirements.txt
flask==3.0.3
flask-mysqldb
python-dotenv
mysqlclient